requirements.txt
requests==2.28.2
beautifulsoup4==4.12.2
discord~=2.1.0
discord-webhook>=1.3.0